Fix Array#divide :bug: :test:
In cases where no match could be made, or where the first element did not match the pattern, `undefined method `<<' for nil:NilClass` would be raised.

Added extra check to add an empty array to memo when the first element did not match the pattern, fixes both problems.

Includes tests and demos